test(worker-sessions): finalize invoke continuation evidence
