+++ docs/internal/baselines/backend-package-file-count.json
@@ -284,6 +284,12 @@
       "removalReason": "pre-existing baseline debt recorded 2026-08-08; split this package into focused sub-packages and remove this entry."
     },
     {
+      "packagePath": "pkg/services/worker_sessions/internal/service",
+      "fileCount": 17,
+      "owner": "worker-maintainers",
+      "removalReason": "Split Worker Sessions continuation and observation responsibilities into focused child packages while preserving the service root contract."
+    },
+    {
       "packagePath": "pkg/services/workers",
       "fileCount": 43,
       "owner": "worker-maintainers",
