+++ pkg/services/worker_sessions/internal/service/invoke_session.go
@@ -2,10 +2,8 @@ package service
 
 import (
 	"context"
-	"encoding/base64"
 	"errors"
 	"fmt"
-	"slices"
 	"sort"
 	"strings"
 	"sync"
@@ -477,90 +475,6 @@ func (r *registry) GetObservationByWorkerSessionID(ctx context.Context, req work
 	}
 	r.logger.Info("worker session observation get by Worker Session", "workerSessionID", projected.WorkerSessionID, "outcome", "success")
 	return projected, nil
-}
-
-func (r *registry) ListWorkerSessionObservations(
-	ctx context.Context,
-	req workersessions.ListWorkerSessionObservationsRequest,
-) (workersessions.ListWorkerSessionObservationsResult, error) {
-	if err := req.Validate(); err != nil {
-		r.logger.Info("worker session top-level observation list rejected", "outcome", "invalid")
-		return workersessions.ListWorkerSessionObservationsResult{}, err
-	}
-	if err := observationContextError(ctx); err != nil {
-		return workersessions.ListWorkerSessionObservationsResult{}, err
-	}
-	limit := req.MaxResults
-	if limit == 0 {
-		limit = workersessions.DefaultWorkerSessionObservationListMaxResults
-	}
-	cursor := ""
-	if strings.TrimSpace(req.NextToken) != "" {
-		decoded, err := base64.StdEncoding.DecodeString(strings.TrimSpace(req.NextToken))
-		if err != nil {
-			return workersessions.ListWorkerSessionObservationsResult{}, workersessions.ErrInvalidObservationPagination
-		}
-		cursor = string(decoded)
-	}
-
-	scope := req.Scope.Normalized()
-	r.mu.RLock()
-	ids := make([]string, 0, len(r.observations))
-	for id, metadata := range r.observations {
-		if metadata == nil || id <= cursor || !observationScopeMatches(metadata.direct, scope) {
-			continue
-		}
-		session, exists := r.sessions[id]
-		if !exists || !observationStateMatches(session.State, req.States) {
-			continue
-		}
-		ids = append(ids, id)
-	}
-	r.mu.RUnlock()
-	sort.Strings(ids)
-
-	pageIDs := ids
-	if len(pageIDs) > limit {
-		pageIDs = pageIDs[:limit]
-	}
-	observations := make([]workersessions.Observation, 0, len(pageIDs))
-	for _, id := range pageIDs {
-		projected, err := r.projectObservation(ctx, id)
-		if err != nil {
-			return workersessions.ListWorkerSessionObservationsResult{}, err
-		}
-		observations = append(observations, projected)
-	}
-	nextToken := ""
-	if len(ids) > len(pageIDs) && len(pageIDs) > 0 {
-		nextToken = base64.StdEncoding.EncodeToString([]byte(pageIDs[len(pageIDs)-1]))
-	}
-	r.logger.Info("worker session top-level observation list", "scope", string(scope), "state_count", len(req.States), "result_count", len(observations), "has_next", nextToken != "")
-	return workersessions.ListWorkerSessionObservationsResult{
-		Observations: observations,
-		MaxResults:   limit,
-		NextToken:    nextToken,
-	}, nil
-}
-
-func observationScopeMatches(direct bool, scope workersessions.ObservationScope) bool {
-	switch scope.Normalized() {
-	case workersessions.ObservationScopeDirect:
-		return direct
-	case workersessions.ObservationScopeFactory:
-		return !direct
-	case workersessions.ObservationScopeAll:
-		return true
-	default:
-		return false
-	}
-}
-
-func observationStateMatches(state workersessions.State, states []workersessions.State) bool {
-	if len(states) == 0 {
-		return true
-	}
-	return slices.Contains(states, state)
 }
 
 func (r *registry) projectObservation(ctx context.Context, id string) (workersessions.Observation, error) {
